Bug fixes, Point, and colors
Users can now add points that automatically fit themselves to a given function. Unfortunately, this means that free-floating points are difficult to create, but this use case is rare because labels fill that need too.

There was a bug in which adding two labels at once incorrectly merged them and broke their reposition code. This bug should be fixed now.

Also added all Khan-Academy-approved colors to the color select menu.

Minor change on realtime.transformY: now the flip around the y-axis to handle how canvas draws is centralized within the transformY method. Previously, everywhere the program drew to canvas it needed a negative sign in front of the y-coordinate.